// apps/sync-service/src/common/infrastructure/database/types/db.d.ts
export interface Logger {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  id: Generated<number>;
  text: Generated<string | null>;
  updated_at: Generated<Date>;
}

export interface MappingActivities {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_activity_id: number;
  existing_program_id: Generated<number | null>;
  id: Generated<number>;
  platform_activity_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingAssetModels {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_asset_model_id: number;
  id: Generated<number>;
  platform_asset_model_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingAssets {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_asset_id: number;
  existing_source_type: Generated<string | null>;
  id: Generated<number>;
  platform_asset_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingAssetTypes {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_asset_type_id: number;
  id: Generated<number>;
  platform_asset_type_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingAssetVendors {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_asset_vendor_id: number;
  id: Generated<number>;
  platform_asset_vendor_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingBatches {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_batch_id: number;
  id: Generated<number>;
  platform_batch_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingBudgetSources {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_budget_source_id: number;
  id: Generated<number>;
  platform_budget_source_id: number;
  platform_global_id: Generated<number | null>;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingCce {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_cce_id: number;
  id: Generated<number>;
  platform_cce_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingCceigat {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_cceigat_id: number;
  id: Generated<number>;
  platform_cceigat_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingColdstorageMaterials {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_coldstorage_material_id: number;
  id: Generated<number>;
  platform_coldstorage_material_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingColdstoragePerTemperatures {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_coldstorage_per_temperature_id: number;
  id: Generated<number>;
  platform_coldstorage_per_temperature_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingColdstorages {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_coldstorage_id: number;
  id: Generated<number>;
  platform_coldstorage_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingCommitmentItems {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_commitment_item_id: number;
  id: Generated<number>;
  platform_commitment_item_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingCommitments {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_commitment_id: number;
  id: Generated<number>;
  platform_commitment_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingCommunicationProviders {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_communication_provider_id: number;
  id: Generated<number>;
  platform_communication_provider_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingContracts {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_contract_id: number;
  id: Generated<number>;
  platform_contract_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingEntities {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_entity_id: number;
  id: Generated<number>;
  platform_entity_id: number;
  platform_global_id: Generated<number | null>;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingEntityMaterialActivities {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_entity_material_activity_id: number;
  existing_entity_material_id: number;
  id: Generated<number>;
  platform_entity_material_activity_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingEventReports {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_event_report_id: number;
  id: Generated<number>;
  platform_event_report_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingExterminationFlowReasons {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_extermination_flow_reason_id: number;
  id: Generated<number>;
  platform_extermination_flow_reason_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingExterminationFlows {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_extermination_flow_id: number;
  id: Generated<number>;
  platform_extermination_flow_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingExterminationShipmentComments {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_extermination_shipment_comment_id: number;
  id: Generated<number>;
  platform_extermination_shipment_comment_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingExterminationShipmentItems {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_extermination_shipment_item_id: number;
  id: Generated<number>;
  platform_extermination_shipment_item_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingExterminationShipments {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_extermination_shipment_id: number;
  id: Generated<number>;
  platform_extermination_shipment_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingExterminationShipmentStocks {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_extermination_shipment_stock_id: number;
  id: Generated<number>;
  platform_extermination_shipment_stock_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingExterminationTransaction {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_extermination_transaction_id: number;
  id: Generated<number>;
  platform_extermination_transaction_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingExterminationTransactionTypes {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_extermination_transaction_type_id: number;
  id: Generated<number>;
  platform_extermination_transaction_type_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingLoggerHistories {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_logger_history_id: number;
  id: Generated<number>;
  platform_logger_history_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingManufactures {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_manufacture_id: number;
  id: Generated<number>;
  platform_global_id: Generated<number | null>;
  platform_manufacture_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingMaterials {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_material_id: number;
  id: Generated<number>;
  platform_global_id: Generated<number | null>;
  platform_material_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingModelRelations {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_model_relation_id: number;
  id: Generated<number>;
  platform_model_relation_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingOrderComments {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_order_comment_id: number;
  id: Generated<number>;
  platform_order_comment_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}

export interface MappingOrderHistories {
  created_at: Generated<Date>;
  deleted_at: Generated<Date | null>;
  existing_order_history_id: number;
  id: Generated<number>;
  platform_order_history_id: number;
  program_id: number;
  updated_at: Generated<Date>;
}
// ...
export interface DB {
  logger: Logger;
  mapping_activities: MappingActivities;
  mapping_asset_models: MappingAssetModels;
  mapping_asset_types: MappingAssetTypes;
  mapping_asset_vendors: MappingAssetVendors;
  mapping_assets: MappingAssets;
  mapping_batches: MappingBatches;
  mapping_budget_sources: MappingBudgetSources;
  mapping_cce: MappingCce;
  mapping_cceigat: MappingCceigat;
  mapping_coldstorage_materials: MappingColdstorageMaterials;
  mapping_coldstorage_per_temperatures: MappingColdstoragePerTemperatures;
  mapping_coldstorages: MappingColdstorages;
  mapping_commitment_items: MappingCommitmentItems;
  mapping_commitments: MappingCommitments;
  mapping_communication_providers: MappingCommunicationProviders;
  mapping_contracts: MappingContracts;
  mapping_entities: MappingEntities;
  mapping_entity_material_activities: MappingEntityMaterialActivities;
  mapping_event_reports: MappingEventReports;
  mapping_extermination_flow_reasons: MappingExterminationFlowReasons;
  mapping_extermination_flows: MappingExterminationFlows;
  mapping_extermination_shipment_comments: MappingExterminationShipmentComments;
  mapping_extermination_shipment_items: MappingExterminationShipmentItems;
  mapping_extermination_shipment_stocks: MappingExterminationShipmentStocks;
  mapping_extermination_shipments: MappingExterminationShipments;
  mapping_extermination_transaction: MappingExterminationTransaction;
  mapping_extermination_transaction_types: MappingExterminationTransactionTypes;
  mapping_logger_histories: MappingLoggerHistories;
  mapping_manufactures: MappingManufactures;
  mapping_materials: MappingMaterials;
  mapping_model_relations: MappingModelRelations;
  mapping_order_comments: MappingOrderComments;
  mapping_order_histories: MappingOrderHistories;
  mapping_order_item_projection_capacities: MappingOrderItemProjectionCapacities;
  mapping_order_items: MappingOrderItems;
  mapping_order_stocks: MappingOrderStocks;
  mapping_orders: MappingOrders;
  mapping_patients: MappingPatients;
  mapping_pqs: MappingPqs;
  mapping_programs: MappingPrograms;
  mapping_reconciliations: MappingReconciliations;
  mapping_stock_exterminations: MappingStockExterminations;
  mapping_stock_opname_periods: MappingStockOpnamePeriods;
  mapping_stock_opnames: MappingStockOpnames;
  mapping_stocks: MappingStocks;
  mapping_transaction_reasons: MappingTransactionReasons;
  mapping_transactions: MappingTransactions;
  mapping_user_changelogs: MappingUserChangelogs;
  mapping_users: MappingUsers;
  mapping_volume_material: MappingVolumeMaterial;
}