// apps/sync-service/src/openapi/order-status-allocate.ts
export type PutV2OrderIdAllocateBodyItem = {
  /** @nullable */
  status: number | null;
  order_item_id: number;
  allocated_stock_id: number;
  allocated_qty: number;
};